Add builtin function to create lists
This is somewhat redundant with the list syntax, but it is more
powerful, while the list syntax is usually way more convenient.

Eventually I want to make the parser extensible, and implement the
then-optional list syntax on top of this new primitive, but for now it's
okay for both of them to exist side by side.