Split `Source` into enum and struct
So far, we act as if sources were always continous but, as the comment
in `merge` documents` they aren't. This commit paves the way for
supporting discontinuous sources and "null sources" (which can be used
in place of a real source, where a source would be expected but none is
available).